changed capacity to roomcount
changed primary key to int

+++ hotelservice/src/main/java/com/example/hotelservice/repositories/HotelRepository.java
@@ -3,5 +3,5 @@ package com.example.hotelservice.repositories;
 import com.example.hotelservice.entity.Hotel;
 import org.springframework.data.r2dbc.repository.R2dbcRepository;
 
-public interface HotelRepository extends R2dbcRepository<Hotel, Long> {
+public interface HotelRepository extends R2dbcRepository<Hotel, Integer> {
 }
