+++ hotelservice/src/main/java/com/example/hotelservice/repositories/LocationRepository.java
@@ -3,5 +3,5 @@ package com.example.hotelservice.repositories;
 import com.example.hotelservice.entity.Location;
 import org.springframework.data.r2dbc.repository.R2dbcRepository;
 
-public interface LocationRepository extends R2dbcRepository<Location, Long> {
+public interface LocationRepository extends R2dbcRepository<Location, Integer> {
 }
